frontend: Segregate EditDialog forms for groups into a seperate dir
